+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.3
+// swift-tools-version:5.9
 
 import PackageDescription
 
@@ -6,9 +6,9 @@ let package = Package(
     name: "PLCrashReporter",
     defaultLocalization: "en",
     platforms: [
-        .iOS(.v11),
-        .macOS(.v10_10),
-        .tvOS(.v11)
+        .iOS(.v12),
+        .macOS(.v10_13),
+        .tvOS(.v12)
     ],
     products: [
         .library(name: "CrashReporter", targets: ["CrashReporter"])
@@ -18,17 +18,8 @@ let package = Package(
             name: "CrashReporter",
             path: "",
             exclude: [
-                "Source/dwarf_opstream.hpp",
-                "Source/dwarf_stack.hpp",
-                "Source/PLCrashAsyncDwarfCFAState.hpp",
-                "Source/PLCrashAsyncDwarfCIE.hpp",
-                "Source/PLCrashAsyncDwarfEncoding.hpp",
-                "Source/PLCrashAsyncDwarfExpression.hpp",
-                "Source/PLCrashAsyncDwarfFDE.hpp",
-                "Source/PLCrashAsyncDwarfPrimitives.hpp",
-                "Source/PLCrashAsyncLinkedList.hpp",
                 "Source/PLCrashReport.proto",
-                "Tools/CrashViewer/",	
+                "Tools/CrashViewer/",
                 "Other Sources/Crash Demo/",
                 "Dependencies/protobuf-c/generate-pb-c.sh",
             ],
@@ -43,6 +34,9 @@ let package = Package(
                 .define("PLCRASHREPORTER_PREFIX", to: ""),
                 .define("SWIFT_PACKAGE"), // Should be defined by default, Xcode 11.1 workaround.
                 .headerSearchPath("Dependencies/protobuf-c")
+            ],
+            swiftSettings: [
+                .interoperabilityMode(.Cxx),
             ],
             linkerSettings: [
                 .linkedFramework("Foundation")